requirements-dev.txt
-r requirements.txt
nodeenv
pydocstyle
pylint
pylint-django
splinter
prospector
mypy
pyroma
djangorestframework-stubs
